chore: move getMediaDirectoryName to ts
