+++ src/hooks/directoryHooks/useUpdateDirectoryHook.jsx
@@ -10,9 +10,10 @@ import {
 
 import { ServicesContext } from "contexts/ServicesContext"
 
+import { getMediaDirectoryName } from "utils/media"
 import { useSuccessToast, useErrorToast } from "utils/toasts"
 
-import { DEFAULT_RETRY_MSG, getMediaDirectoryName } from "utils"
+import { DEFAULT_RETRY_MSG } from "utils"
 
 import { extractUpdateDirectoryInfo } from "./utils"
