+++ src/utils/media.ts
@@ -33,3 +33,32 @@ export const getSelectedMediaDto = (fileData: MediaData) => {
 
   return selectedData
 }
+
+interface GetMediaDirectoryNameSplitProps {
+  start?: number
+  end?: number
+  splitOn?: string
+  joinOn?: string
+  decode?: boolean
+}
+
+export const getMediaDirectoryName = (
+  mediaDirectoryName: string,
+  {
+    start = 0,
+    end,
+    splitOn = "%2F",
+    joinOn = "%2F",
+    decode = false,
+  }: GetMediaDirectoryNameSplitProps
+) => {
+  const mediaDirectoryArray = mediaDirectoryName.split(splitOn)
+  const selectedMediaDirectoryArray = mediaDirectoryArray.slice(start, end)
+  if (decode) {
+    const decodedSelectedMediaDirectoryArray = selectedMediaDirectoryArray.map(
+      (v) => decodeURIComponent(v)
+    )
+    return decodedSelectedMediaDirectoryArray.join(joinOn)
+  }
+  return selectedMediaDirectoryArray.join(joinOn)
+}
